backend/pyproject.toml
[tool.poetry]
name = "mis_eventos_backend"
version = "0.1.0"
description = "Backend para la gestión de eventos usando Flask"
authors = ["Libardo bejaranolibardo5@gmail.com"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0"
sqlalchemy = "^2.0"
sqlmodel = "^0.0.16"
psycopg2-binary = "^2.9"
alembic = "^1.13"
python-dotenv = "^1.0"
flask-sqlalchemy = "^3.1.1"
marshmallow = "^4.0.0"
pyjwt = "^2.10.1"
flask-restx = "^1.3.0"
gunicorn = "^23.0.0"
flask-cors = "^5.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

